adapt issues from re-basing to develop
* changes form current develop: 20ef2075f1a9d9836e6231446e0e8ff7b21f5b93
