+++ src/main/java/de/uhd/ifi/se/decision/management/jira/ComponentGetter.java
@@ -15,9 +15,11 @@ import com.atlassian.sal.api.user.UserManager;
 import de.uhd.ifi.se.decision.management.jira.config.AuthenticationManager;
 import de.uhd.ifi.se.decision.management.jira.persistence.ActiveObjectPersistenceManager;
 import de.uhd.ifi.se.decision.management.jira.persistence.JiraIssueTextPersistenceManager;
-
 import de.uhd.ifi.se.decision.management.jira.extraction.GitExtractor;
-import org.eclipse.jgit.api.Git;
+
+
+import java.util.HashMap;
+import java.util.Map;
 
 /**
  * @description Provides access to JIRA components. Automatically initialized.
@@ -41,6 +43,17 @@ public class ComponentGetter {
 	private static UserManager userManager;
 	@ComponentImport
 	private static ActiveObjects activeObjects;
+
+	/**
+	 * @issue How to access knowledge extracted from commits?
+	 * @decision Hold cached commits and their knowledge in globaly accessible class.
+	 * @pro consistency: git repositories are another knowledge source like activeObjects or issueService
+	 * @con resources: consumes more memory, risk of resource starving with poor implementation
+	 * @alternative instantiate classes for commit knowledge extraction only when needed.
+	 * @con performance: IO access is costly and requires more CPU time
+	 * @pro resources: consumes less memory, CPU time costs are low
+	 */
+	private static Map<String, GitExtractor> gitExtractors = null; // projecShortName as the key
 
 	@Inject
 	public ComponentGetter(UserManager userManager, ActiveObjects activeObjects) {
