+++ src/test/java/de/uhd/ifi/se/decision/management/jira/quality/commentmetriccalculator/TestGetNumberOfRelevantSentencesComment.java
@@ -2,19 +2,15 @@ package de.uhd.ifi.se.decision.management.jira.quality.commentmetriccalculator;
 
 import static org.junit.Assert.assertEquals;
 
-import java.util.Map;
-
 import org.junit.Test;
 
 import net.java.ao.test.jdbc.NonTransactional;
 
-public class GetNumberOfCommentsForJiraIssuesComment extends SetupCommentCalculator {
+public class TestGetNumberOfRelevantSentencesComment extends SetupCommentCalculator {
 
 	@Test
 	@NonTransactional
 	public void testCase() {
-		Map<String, Integer> map = calculator.getNumberOfCommentsForJiraIssues();
-		// TODO this should be 1
-		assertEquals(0, map.size(), 0.0);
+		assertEquals(2, calculator.getNumberOfRelevantSentences().size(), 0.0);
 	}
 }
