+++ src/test/java/de/uhd/ifi/se/decision/management/jira/quality/commonmetriccalculator/TestGetDecKnowlElementsOfATypeGroupedByHavingElementsOfOtherTypeCommon.java
@@ -1,22 +1,15 @@
 package de.uhd.ifi.se.decision.management.jira.quality.commonmetriccalculator;
 
-import de.uhd.ifi.se.decision.management.jira.TestSetUpWithIssues;
+
 import de.uhd.ifi.se.decision.management.jira.model.KnowledgeType;
-import de.uhd.ifi.se.decision.management.jira.persistence.AbstractPersistenceManager;
-import net.java.ao.test.jdbc.Data;
-import net.java.ao.test.jdbc.NonTransactional;
-import net.java.ao.test.junit.ActiveObjectsJUnitRunner;
 import org.junit.Ignore;
 import org.junit.Test;
-import org.junit.runner.RunWith;
 
 import java.util.Map;
 
 import static org.junit.Assert.*;
 
-@RunWith(ActiveObjectsJUnitRunner.class)
-@Data(TestSetUpWithIssues.AoSentenceTestDatabaseUpdater.class)
-public class TestGetDecKnowlElementsOfATypeGroupedByHavingElementsOfOtherType extends TestSetupCalculator {
+public class TestGetDecKnowlElementsOfATypeGroupedByHavingElementsOfOtherTypeCommon extends SetupCommonCalculator {
 	@Test
 	public void testTypesNull() {
 		assertEquals(0, calculator.getDecKnowlElementsOfATypeGroupedByHavingElementsOfOtherType(null, null).size(), 0.0);
@@ -25,13 +18,11 @@ public class TestGetDecKnowlElementsOfATypeGroupedByHavingElementsOfOtherType ex
 	}
 
 	@Test
-	@NonTransactional
 	public void testSameTypes() {
 		assertEquals(0, calculator.getDecKnowlElementsOfATypeGroupedByHavingElementsOfOtherType( KnowledgeType.DECISION,KnowledgeType.DECISION).size(), 0.0);
 	}
 
 	@Test
-	@NonTransactional
 	public void testForDifferentTypesNotLinked() {
 		Map<String, String> calculation = calculator.getDecKnowlElementsOfATypeGroupedByHavingElementsOfOtherType(
 				argumentElement.getType(), decisionElement.getType());
@@ -53,17 +44,16 @@ public class TestGetDecKnowlElementsOfATypeGroupedByHavingElementsOfOtherType ex
 	}
 
 	@Test
-	@NonTransactional
 	@Ignore
 	public void testForDifferentTypesLinked() {
 		//long linkId = linkElements(user, decisionElement, argumentElement, "");
-
+/*
 		Map<String, String> calculation = calculator.getDecKnowlElementsOfATypeGroupedByHavingElementsOfOtherType(
 				argumentElement.getType(), decisionElement.getType());
 
 		Map<String, String> calculation2 = calculator.getDecKnowlElementsOfATypeGroupedByHavingElementsOfOtherType(
 				decisionElement.getType(), argumentElement.getType());
-/*
+
 		assertEquals(2, calculation.size(),0.0); // expecting always two categories
 		assertTrue(calculation.containsKey("Argument has no Decision"));
 		assertEquals("", calculation.get("Argument has no Decision"));
