+++ src/test/java/de/uhd/ifi/se/decision/management/jira/quality/commonmetriccalculator/TestGetLinksToIssueTypeMapCommon.java
@@ -7,18 +7,15 @@ import org.junit.Ignore;
 import org.junit.Test;
 
 import de.uhd.ifi.se.decision.management.jira.model.KnowledgeType;
-import net.java.ao.test.jdbc.NonTransactional;
 
-public class TestGetLinksToIssueTypeMap extends TestSetupCalculator {
+public class TestGetLinksToIssueTypeMapCommon extends SetupCommonCalculator {
 
 	@Test
-	@NonTransactional
 	public void testTypeNull() {
 		assertNull(calculator.getLinksToIssueTypeMap(null));
 	}
 
 	@Test
-	@NonTransactional
 	@Ignore
 	public void testTypeFilled() {
 		Object map = calculator.getLinksToIssueTypeMap(KnowledgeType.ARGUMENT);
