+++ src/test/java/de/uhd/ifi/se/decision/management/jira/quality/commonmetriccalculator/TestGetNumberOfLinkToOtherElementCommon.java
@@ -5,33 +5,28 @@ import static org.junit.Assert.assertEquals;
 import org.junit.Test;
 
 import de.uhd.ifi.se.decision.management.jira.model.KnowledgeType;
-import net.java.ao.test.jdbc.NonTransactional;
 
-public class TestGetNumberOfLinkToOtherElement extends TestSetupCalculator {
+public class TestGetNumberOfLinkToOtherElementCommon extends SetupCommonCalculator {
 
 	@Test
-	@NonTransactional
 	public void testLinkFromNullLinkToNull() {
 		assertEquals(0, calculator._getNumberOfLinksToOtherElement(
 				null, null).size(), 0.0);
 	}
 
 	@Test
-	@NonTransactional
 	public void testLinkFromFilledLinkToNull() {
 		assertEquals(0, calculator._getNumberOfLinksToOtherElement(
 				KnowledgeType.DECISION, null).size(), 0.0);
 	}
 
 	@Test
-	@NonTransactional
 	public void testLinkFromNullLinkToFilled() {
 		assertEquals(0, calculator._getNumberOfLinksToOtherElement(
 				null, KnowledgeType.ISSUE).size(), 0.0);
 	}
 
 	@Test
-	@NonTransactional
 	public void testLinkFromFilledLinkToFilled() {
 		assertEquals(2, calculator._getNumberOfLinksToOtherElement(
 				KnowledgeType.ARGUMENT, KnowledgeType.DECISION).size(),
